raw conversion of AsyncAwait specs
